Refactor token index into a new directory
